Basic support for UL ODISYS log
Use VS 2013 community ed.
Handle all uncaught exceptions
Support UL ODISYS logs (in PlainTextLogFormat)
Add new line column for Thread
Change default time stamp display to long
Search also on Thread column
Speedup file load by removing usage to ConcurrentDictionnary.Count
Remove UDP source
Start center in screen, with larger size

+++ LogWatch/Features/Records/TimestampFormat.cs
@@ -1,6 +1,6 @@
 ﻿namespace LogWatch.Features.Records {
     public enum TimestampFormat {
-        Short,
-        Long
+        Long,
+        Short
     }
 }
